Basic construction
The basic functions including product selected by the buyer, and send
email to buyer and seller after purchasing are basically constructed.
For the product selected by the buyer, it is a dummy data that is not
come form session. Session function will be added. User interface and
error handling will be improved.